iOS - SupportKit to Smooch rename
