+++ README.md
@@ -1,64 +1,64 @@
-# supportkit-cordova
-Cordova (PhoneGap) bindings for [SupportKit](https://supportkit.io).
+# smooch-cordova
+Cordova (PhoneGap) bindings for [Smooch](https://smooch.io).
 
 # Basic Setup
 
-## Install SupportKit
+## Install Smooch
 
-Install SupportKit iOS or Android libraries,
+Install Smooch iOS or Android libraries,
 
-On iOS: [http://docs.supportkit.io/#manual-method](http://docs.supportkit.io/#manual-method)
+On iOS: [http://docs.smooch.io/#manual-method](http://docs.smooch.io/#manual-method)
 
-On Android: [http://docs.supportkit.io/android/#adding-supportkit-to-your-app](http://docs.supportkit.io/android/#adding-supportkit-to-your-app)
+On Android: [http://docs.smooch.io/android/#adding-smooch-to-your-app](http://docs.smooch.io/android/#adding-smooch-to-your-app)
 
 ## Install Cordova plugin
 
-	cordova plugin add io-supportkit-sktcordovaplugin
+	cordova plugin add io-smooch-smcordovaplugin
 
-You should now see supportkit-cordova.js and the SupportKitCordova source files in your project directory.
+You should now see smooch-cordova.js and the SmoochCordova source files in your project directory.
 
-## Init SupportKit with appToken
+## Init Smooch with appToken
 
 ### iOS
 
 In your javascript:
 
-	SupportKit.init({
-    	"appToken" : "your-supportkit-app-token",
+	Smooch.init({
+    	"appToken" : "your-smooch-app-token",
     	"conversationAccentColor" : "#FFFFFF" // optional, default is #00B0FF
 	});
 	
 **IMPORTANT**: This call must be made **AFTER** the `deviceready` event of Cordova.
 
-See [http://docs.supportkit.io/api/Classes/SKTSettings.html](http://docs.supportkit.io/api/Classes/SKTSettings.html) for other init settings.
+See [http://docs.smooch.io/api/Classes/SKTSettings.html](http://docs.smooch.io/api/Classes/SKTSettings.html) for other init settings.
 
 ### Android
 
-Init SupportKit(natively) in the Application class, for example:
+Init Smooch(natively) in the Application class, for example:
 
 	public class MainApplication extends Application {
 	    public void onCreate() {
 	        super.onCreate();
-	        SupportKit.init(this, "your-supportkit-app-token");
+	        Smooch.init(this, "your-smooch-app-token");
 	    }
 	}
 
 The javascript init call will not work in Android. It must be done natively.
 
-See [http://docs.supportkit.io/api/android/](http://docs.supportkit.io/api/android/) for other init settings.
+See [http://docs.smooch.io/api/android/](http://docs.smooch.io/api/android/) for other init settings.
 
 
-## Show SupportKit
+## Show Smooch
 
-Now, you're ready to show the SupportKit chat window.
+Now, you're ready to show the Smooch chat window.
 
-	SupportKit.show();
+	Smooch.show();
 
 # Awesome Setup
 
 ## Add user info
 
-	SupportKit.setUser({
+	Smooch.setUser({
 		"firstName" : "Erlich",
 		"lastName" : "Bachman",
 		"email" : "erlich.bachman@aviato.com",
@@ -67,20 +67,12 @@ Now, you're ready to show the SupportKit chat window.
 
 ## Add custom user properties
 
-	SupportKit.setUserProperties({
+	Smooch.setUserProperties({
 		"tv_show" : "Silicon Valley",
 		"episodes" : "Server Space, The Lady, Bad Money",
 		"net_worth" : 1000000
 	});
 
-## Track events to show SupportKit Whispers
+## Track events to show Smooch Whispers
 
-    SupportKit.track("Slapped kid");
-
-## Set Default Recommended Articles (iOS only)
-	
-	SupportKit.setDefaultRecommendations(["http://m.imdb.com/name/nm3042755", "http://m.imdb.com/name/nm2554352"])
-
-## Set Top Recommended Article (iOS only)
-
-	SupportKit.setTopRecommendation("https://supportkit.zendesk.com/hc/en-us/articles/201900704-Giving-your-users-the-right-answer-at-the-right-time")
+    Smooch.track("Slapped kid");
