+++ package.json
@@ -1,10 +1,10 @@
 {
-  "version": "1.1.3",
-  "name": "io-supportkit-sktcordovaplugin",
-  "cordova_name": "SupportKit",
-  "description": "SupportKit Cordova Plugin for iOS and Android",
+  "version": "1.1.4",
+  "name": "io-smooch-smcordovaplugin",
+  "cordova_name": "Smooch",
+  "description": "Smooch Cordova Plugin for iOS and Android",
   "keywords": [
-    "SupportKit",
+    "Smooch",
     "messaging",
     "chat",
     "in-app",
@@ -18,7 +18,7 @@
   ],
   "engines": [],
   "cordova": {
-    "id": "io-supportkit-sktcordovaplugin",
+    "id": "io-smooch-smcordovaplugin",
     "platforms": [
       "ios",
       "android"
@@ -26,12 +26,12 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/supportkit/supportkit-cordova.git"
+    "url": "git+https://github.com/smooch/smooch-cordova.git"
   },
-  "author": "SupportKit",
+  "author": "Smooch",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/supportkit/supportkit-cordova/issues"
+    "url": "https://github.com/smooch/smooch-cordova/issues"
   },
-  "homepage": "https://github.com/supportkit/supportkit-cordova#readme"
+  "homepage": "https://github.com/smooch/smooch-cordova#readme"
 }
